+++ pcs_test/tier0/cli/test_resource.py
@@ -461,13 +461,13 @@ class ResourceMoveBanMixin:
         )
 
 
-class ResourceMove(ResourceMoveBanMixin, TestCase):
+class ResourceMoveLegacy(ResourceMoveBanMixin, TestCase):
     def setUp(self):
         self.lib = mock.Mock(spec_set=["resource"])
         self.resource = mock.Mock(spec_set=["move"])
         self.lib.resource = self.resource
         self.lib_command = self.resource.move
-        self.cli_command = resource.resource_move
+        self.cli_command = resource.resource_move_with_constraint
         self.no_args_error = "must specify a resource to move"
 
 
@@ -479,6 +479,86 @@ class ResourceBan(ResourceMoveBanMixin, TestCase):
         self.lib_command = self.resource.ban
         self.cli_command = resource.resource_ban
         self.no_args_error = "must specify a resource to ban"
+
+
+@mock.patch("pcs.resource.warn")
+class ResourceMove(TestCase):
+    def setUp(self):
+        self.lib = mock.Mock(spec_set=["resource"])
+        self.resource = mock.Mock(spec_set=["move_autoclean"])
+        self.lib.resource = self.resource
+
+    def test_no_args(self, mock_warn):
+        with self.assertRaises(CmdLineInputError) as cm:
+            resource.resource_move(self.lib, [], dict_to_modifiers(dict()))
+        self.assertEqual(
+            cm.exception.message, "must specify a resource to move"
+        )
+        self.resource.move_autoclean.assert_not_called()
+        mock_warn.assert_not_called()
+
+    def test_too_many_args(self, mock_warn):
+        with self.assertRaises(CmdLineInputError) as cm:
+            resource.resource_move(
+                self.lib,
+                ["resource", "arg1", "arg2"],
+                dict_to_modifiers(dict()),
+            )
+        self.assertIsNone(cm.exception.message)
+        self.resource.move_autoclean.assert_not_called()
+        mock_warn.assert_not_called()
+
+    def test_succes(self, mock_warn):
+        resource.resource_move(
+            self.lib, ["resource"], dict_to_modifiers(dict())
+        )
+        self.resource.move_autoclean.assert_called_once_with(
+            "resource", node=None, master=False, wait_timeout=-1, strict=False
+        )
+        mock_warn.assert_not_called()
+
+    def test_success_node(self, mock_warn):
+        resource.resource_move(
+            self.lib, ["resource", "node"], dict_to_modifiers(dict())
+        )
+        self.resource.move_autoclean.assert_called_once_with(
+            "resource", node="node", master=False, wait_timeout=-1, strict=False
+        )
+        mock_warn.assert_not_called()
+
+    def test_success_wait(self, mock_warn):
+        resource.resource_move(
+            self.lib, ["resource", "node"], dict_to_modifiers(dict(wait=None))
+        )
+        self.resource.move_autoclean.assert_called_once_with(
+            "resource", node="node", master=False, wait_timeout=0, strict=False
+        )
+        mock_warn.assert_not_called()
+
+    def test_success_autodelete(self, mock_warn):
+        resource.resource_move(
+            self.lib,
+            ["resource", "node"],
+            dict_to_modifiers(dict(autodelete=True)),
+        )
+        self.resource.move_autoclean.assert_called_once_with(
+            "resource", node="node", master=False, wait_timeout=-1, strict=False
+        )
+        mock_warn.assert_called_once_with(
+            "Option '--autodelete' is deprecated. There is no need to use it "
+            "as its functionallity is default now."
+        )
+
+    def test_success_all_options(self, mock_warn):
+        resource.resource_move(
+            self.lib,
+            ["resource", "node"],
+            dict_to_modifiers(dict(master=True, strict=True, wait="10")),
+        )
+        self.resource.move_autoclean.assert_called_once_with(
+            "resource", node="node", master=True, wait_timeout=10, strict=True
+        )
+        mock_warn.assert_not_called()
 
 
 class ResourceClear(TestCase):